Prevent unintended slowdowns
The homing process calculates the moving coordinates several times based on the current sys_positon. When setting sys_position to 0 in advance for homing, it is necessary to set not only the radius axis but also the polar axis coordinates to 0.

In case the axis homed alone, like $HY and $HZ, post homing also needs to be processed by determining the mask value, so I added the mask value to the function argument.

When recalculating the movement speed, the distance traveled by the Polar axis is now approximated by the ratio of the arc length instead of the change in angle.

Fixed a problem where the speed multiplier was multiplying the result of the previous calculation.